refactor: mv to helpers.go
